+++ dynaml-core/src/main/scala-2.11/io/github/mandar2812/dynaml/evaluation/RegressionMetricsTF.scala
@@ -19,8 +19,7 @@ under the License.
 package io.github.mandar2812.dynaml.evaluation
 
 import com.quantifind.charts.Highcharts.{regression, title, xAxis, yAxis}
-import io.github.mandar2812.dynaml.tensorflow.dtf
-import io.github.mandar2812.dynaml.tensorflow.utils.dtfutils
+import io.github.mandar2812.dynaml.tensorflow.{dtf, dtfutils}
 import org.platanios.tensorflow.api.{::, Tensor}
 
 /**
